代码完成，准备开始实验

+++ BFC-EM/BFC-EM.h
@@ -2,5 +2,5 @@
 #define bfcem
 #include "../graph.h"
 
-int bfcEm(graph1& g, int maxBufferSize);
+int bfcEm(graph1& g, long long maxBufferSize);
 #endif
